Move create non concerns case
