+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/CreateCase/NonConcernsCase/CreateNonConcernsCase.cshtml
@@ -12,11 +12,11 @@
 
 	@if (currentStep != "2" )
 	{
-		<partial name="_SelectTrust"/>
+		<partial name=""/>
 	}
 	@if (currentStep == "2")
 	{
-		<partial name="_SelectCaseType"/>
+		<partial name=""/>
 	}
 
 </div>
